graphql/api/requirements.txt
alembic==1.8.1
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
fastapi==0.85.2
graphql-core==3.2.3
greenlet==2.0.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
strawberry-graphql==0.138.1
typing-extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
